Refactored React code for separation of concerns
